Fix: constructor parameters were incorrect

+++ src/entities/Archers/Archer.java
@@ -6,7 +6,7 @@ public abstract class Archer extends Character {
     protected boolean visionActive; // Whether vision skill is active or not.
 
     public Archer(int position, int damage, int health, int maxHealth, double defense, int movementSpeed,
-            int range, double critRate, int critDmg) {
+            int range, double critRate, double critDmg) {
         super(position, damage, health, maxHealth, defense, movementSpeed, range, critRate, critDmg);
         visionActive = false;
     }
